backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-genai==0.3.0
python-dotenv==1.0.1
pydantic==2.6.0
ansible-runner==2.3.4
